# src/core/Snakk.Sdk/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Snakk.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:17100/"
    }
  ],
  "paths": {
...
    "/admin/security/audit-logs": {
      "get": {
        "tags": [
          "Admin - Security & Audit"
        ],
        "operationId": "AdminGetAuditLogs",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorUserId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/security/audit-logs/{id}": {
      "get": {
        "tags": [
          "Admin - Security & Audit"
        ],
        "operationId": "AdminGetAuditLog",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogDto"
                }
              }
            }
          }
        }
      }
    },
    "/admin/security/failed-logins": {
      "get": {
        "tags": [
          "Admin - Security & Audit"
        ],
        "operationId": "AdminGetFailedLogins",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "hours",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailedLoginsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/security/suspicious-activities": {
      "get": {
        "tags": [
          "Admin - Security & Audit"
        ],
        "operationId": "AdminGetSuspiciousActivities",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuspiciousActivitiesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/security/user-data-export/{userId}": {
      "post": {
        "tags": [
          "Admin - Security & Audit"
        ],
        "operationId": "AdminExportUserData",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDataExportDto"
                }
              }
            }
          }
        }
      }
    },
    "/admin/security/user-data-export/{exportId}": {
      "get": {
        "tags": [
          "Admin - Security & Audit"
        ],
        "operationId": "AdminGetUserDataExport",
        "parameters": [
          {
            "name": "exportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDataExportStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditLogDto": {
        "required": [
          "id",
          "action",
          "category",
          "severity",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "actorUsername": {
            "type": [
              "null",
              "string"
            ]
          },
          "action": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "targetType": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetId": {
            "type": [
              "null",
              "string"
            ]
          },
          "details": {
            "type": [
              "null",
              "string"
            ]
          },
          "ipAddress": {
            "type": [
              "null",
              "string"
            ]
          },
          "success": {
            "type": "boolean"
          },
          "severity": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuditLogsResponse": {
        "required": [
          "logs",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLogDto"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "FailedLoginsResponse": {
        "required": [
          "failures",
          "suspiciousIps",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "failures": {
            "type": "array"
          },
          "suspiciousIps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuspiciousIpResponse"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "SuspiciousActivitiesResponse": {
        "required": [
          "activities",
          "total"
        ],
        "type": "object",
        "properties": {
          "activities": {
            "type": "array"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "SuspiciousIpResponse": {
        "required": [
          "ipAddress",
          "failureCount",
          "lastAttempt"
        ],
        "type": "object",
        "properties": {
          "ipAddress": {
            "type": "string"
          },
          "failureCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "lastAttempt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UserDataExportDto": {
        "required": [
          "exportId",
          "message"
        ],
        "type": "object",
        "properties": {
          "exportId": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "UserDataExportStatusResponse": {
        "required": [
          "exportId",
          "status",
          "message"
        ],
        "type": "object",
        "properties": {
          "exportId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}